Disable if_modified_since regardless of ETag patch
It seems that just writing a better ETag isn't enough and we still need
to disable If-Modified-Since and Last-Modified, because after the last
update my browser (Firefox 66) has kept some older assets, which in turn
broke the client.

I still need to investigate on this further, but we might need to
disable Last-Modified for store paths in nixpkgs if the cache hit really
was because of that.

Signed-off-by: aszlig <aszlig@nix.build>