Restock Armoire if new Armoire items were added
The Enchanted Armoire regularily gets new items in new upstream
versions, which works fine if users didn't already empty the items that
are available so far.

Unfortunately, if a user has emptied the Armoire, an armoireEmpty flag
is set for the user's record and with subsequent updates the user won't
get new items out of that chest if that flag is set.

On the upstream system, a migration script that restocks the Armoire is
run whenever there are new items, which works fine for them but we
simply do not want to notify users about running that migration script
every time there is an update with new Armoire items.

So instead, we now hash the contents of the Armoire from the source tree
and compare that hash with the one we have in
/var/lib/shabitica/armoire.sha256 and if it differs from the hash
determined from the sources, we simply run the migration script.

Signed-off-by: aszlig <aszlig@nix.build>