Update to version 4.99.1
API changes:

  * Feature: HTTP basic auth middleware now accepts comma-delimited list
             of usernames and passwords

Client changes:

  * Fix: Add chevrons to menus on site when at mobile scale
  * Fix: Correct color of Instagram social media button
  * Fix: Correctly restrict display of Guilds suggested for new accounts
         to <60 days age

Repo changes:

  * Chore: Update several packages per Greenkeeper: validator,
           merge-stream, gulp-imagemin, axios

Rebasing the patches went mostly smooth, but with a few conflicts
because of the changes in basic authenticaton, where we only need to
correct a few variables in the config and of course there were conflicts
in the package.json file because of the Greenkeeper updates.

Another conflict was in the avatar PNG generation, but we didn't use
that in the first place, because it requires Amazon S3. So fixing that
conflict was just a matter of deleting everything.

Moving along with the changes, I updated all of our dependencies and now
pageres also requires puppeteer, so I had to add
PUPPETEER_SKIP_CHROMIUM_DOWNLOAD to pageres as well.

Signed-off-by: aszlig <aszlig@nix.build>