Revert "Update to version 4.99.1"
This reverts commit 870647ba045d145f251f4d03824d67ae06421737.

The reason is that this version has serious UI issues, with tasks not
being editable and the logo sometimes not being displayed correctly.

After deploying this on my server I almost immediately rolled it back
because I wasn't able to save some tasks. So I'm now waiting until
either I get the time to investigate this further (unfortunately, very
unlikely to happen at the moment :-/) or it eventually gets sorted out
upstream.

Signed-off-by: aszlig <aszlig@nix.build>