Update to version 4.100.0
This is the second attempt to bring in the changes version 4.99.1, which
I reverted in d244ee932576dafc940547566afbe6f84c418e04 because of some
issues.

I decided to at least roughly figure out which dependency was causing
the issues and I found out that the bootstrap-vue update to version
2.0.0-rc.22 was the issue.

UI issues with new versions bootstrap-vue are quite frequent in the
development history of Habitica, so this was actually my first guess, so
I decided to actually introduce an exact match on the version for
bootstrap-vue in order to avoid the need to revert on our side in the
future.

Alongside the changes introduced in version 4.99.1, which I summarized
in 870647ba045d145f251f4d03824d67ae06421737, here are the changes for
version 4.100.0:

  * Feature: Send reminder emails to subscribers using Apple/Google
             payment methods when payment is about to come due
  * Content: New pet quest, the Dolphin of Doubt
  * Content: New achievements: All Your Base, Back to Basics, Just Add
             Water, Mind Over Matter
  * Refactor: Lost Masterclasser achievement abstracted to use the same
              logic as new quest achievements
  * Chore: Re-enable Oddballs Quest Bundle
  * Chore: Bailey news announcement re the above
  * Fix: Use HTTPS instead of HTTP for links to moderator contact form

Signed-off-by: aszlig <aszlig@nix.build>