Update to version 4.100.1
Upstream changes:

  * Use pagination feature for private messages
  * Fix: More reasonable quest reward amounts for Dolphin
  * Fix: Correct end date on Oddballs Quest Bundle
  * Fix: Various backgrounds resized and corrected to line up better on
         3x3 pixel grid
  * Chore: Update packages per Greenkeeper: chromedriver, js2xmlparser,
           pageres

A note about the last item: While the official release notes mention
this, the actual diff contains no changes in package.json,
package-lock.json or anything else that might indicate such an update,
so we don't need to take action on our side.

The pagination change introduced a few conflicts on our side, which is
mainly because parts of functionality were moved to server/libs/inbox.

I've subsequently added that new function (sentMessage) from
inbox/index.js to our list of function canaries, so that we get an error
once a new call location for that function is introduced upstream.

Signed-off-by: aszlig <aszlig@nix.build>