migrations: Add migration for Orca pets and mounts
This was actually introduced in version 4.101.0, but this has slipped
through without me noticing because while I did look at the diff, I
later forgot to add that migration.

The migration essentially awards every user logged in after 2019-05-18
an Orca mount. The Orca *pets* are only awarded if the user already has
the mount, but for S*h*abitica this isn't the case, so I guess that will
be the case next year.

Signed-off-by: aszlig <aszlig@nix.build>