tests/e2e: Wait for successful Selenium startup
Sometimes if the test machine is on high load, ChromeDriver isn't ready
yet to accept connections and usually the first test will fail as a
consequence of that.

In order to cope with this, I just added another loop which polls the
status[1] endpoint until the "ready" property becomes true.

[1]: https://w3c.github.io/webdriver/#status

Signed-off-by: aszlig <aszlig@nix.build>