Update to version 4.104.2
Upstream changes:

  * Fix: Move password reset email functionality to our email server
         instead of Gmail due to issues with multi-factor authentication
  * Fix quests migration

The first part actually has the advantage that we now only have one way
to send mail to the mailer daemon, which is by using the sendTxnEmail
functionality, so on our side we suddenly have less code because with
this all the parts in the mailer daemon that deal with "simple" mails
are now gone.

Of course, this also adds a corresponding template to the mailer daemon,
so that we actually get a mail with useful content, rather than the
fallback template.

Signed-off-by: aszlig <aszlig@nix.build>