release.nix: Fix build of manual on NixOS Unstable
A recent change[1] that has been merged to nixpkgs master has factored
out the generation of the options in (among others like JSON) DocBook
format into pkgs.nixosOptionsDoc and subsequently also changed the path
of the XSL files.

Ideally, the generation of the manual on our side would also just use a
single function that is provided via nixpkgs for generating the *full*
manual. The nixosOptionsDoc function however provides *only* the option
list in DocBook format, which is useful to *some* extent but doesn't
deduplicate the extra params we give to DocBook (eg. style sheets for
syntax highlighting).

This and the fact that I only have limited time at the moment is why I
left all the other crap we duplicate on our side as-is and only change
the path to the XSL files if we have a nixosOptionsDoc attribute in
pkgs.

[1]: https://github.com/NixOS/nixpkgs/commit/5cfd034af0afe55e4d25748ec986c71b3bfbe3dd

Signed-off-by: aszlig <aszlig@nix.build>