Update to version 4.109.0
Version 4.107.2 and 4.107.3 were only Bailey news updates and version
4.107.1 only added target="_blank" links to the Tavern and removed the
coveralls dependency.

The latter actually is an external dependency, which I didn't catch when
doing my initial analysis of the dependencies. Fortunately it's only
used in tests (and only via a specific flag we don't use) and not in
production.

So since version 4.107.2 to 4.107.3 didn't contain changes that affect
us very much, I decided it would not be worth updating to and do the
next significant update.

Version 4.108.0 also was just a content update, introducing the new
"Footloose Faun" Mystery Set.

However, version 4.109.0 really has the big changes here:

  API:

    * Feature: quests is now an allowable type for the
               /purchase-hourglass call and can take a quantity value in
               the request body (defaults to 1 if unspecified)
    * Content: New quest available in the Time Travelers Shop,
               "Mysterious Metal Marvels"
    * Fix: System messages in chat don't add spurious additional
           backticks to already escaped text
    * Fix: Individual subscription obtained from a Group Plan is no
           longer canceled if user leaves an unrelated Guild
    * Fix: Push notifications are sent in recipient's language instead
           of sender's
    * Fix: API call to open Mystery Item box includes translated name of
           the item received
    * Chore: Locale files updated
    * Chore: Bailey news announcement of Time Travel quest and August
             last chance

  Client:

    * Feature: Demo button to allow user to preview sound effect theme
               before selecting it
    * Fix: 404 page correctly links to Report a Bug Guild
    * Fix: Usernames in Hall of Heroes are clickable

  Repo:

    * Chore: Node version upgraded to 12

Rebasing our patches involved a few conflicts:

  * Since the push notification stuff now uses the recipient language,
    we hit a few conflicts, but since we remove push notifications
    anyway, it's not a big deal.
  * The target="_blank" change for the Tavern links resulted in a few
    conflicts, but again, not a big deal.
  * For the group subscription change, there were a few new tests, but
    our "dont-cancel-group-subscription-on-leave" patch renders them
    obsolete, so I removed them altogether.
  * With the Time Travelers Shop update, the buy() and
    hourglassPurchase() now got an additional argument, but since the
    analytics object (which we remove) is also passed to these
    functions, so we got a few conflicts.

For the function signature changes regarding the Time Travelers Shop, I
accordingly changed our function canaries to cope with the additional
argument.

Since the Node version got upgraded to version 12, the package-lock.json
got updated as well with a bunch of new versions, whereas the updates of
the direct dependencies are:

  mongoose: 5.5.15 -> 5.6.11

    Looking through the changes, the update seems to be mostly minor as
    far as Habitica is concerned and doesn't have any breaking changes.

  chromedriver: 73.0.0 -> 76.0.1

    We only use this during Nightwatch tests and the changes are mostly
    version bumps regarding the chromedriver upstream project.

  bootstrap-vue: 2.0.0-rc.18 -> 2.0.0-rc.28

    In Habitica version 4.109.0 there was a fix (not documented in the
    release notes) for the task editing issue, so we can update to the
    latest version and also drop the "pin-bootstrap-vue" patch.

    The upstream changes are a bit too big to summarise here, but you
    can find them at https://bootstrap-vue.js.org/docs/misc/changelog/.

    From looking through them, there seem to be no changes that could
    possibly break (S)Habitica.

  svg-url-loader: 2.3.2 -> 3.0.0

    The breaking change is that the "noquotes" got removed, but quickly
    grepping the Habitica source tree reveals that it isn't used.

The coveralls and lcov-result-merger dependencies got removed, so as
initially mentioned we didn't actually use it anyway but it's a good
thing to have a few less dependenies.

Additionally, a new direct dependency ("regenerator-runtime") got
introduced in version 4.109.0 as part of the Node update. The dependency
was already there as a direct dependency of things like babel and pug,
but I haven't found anything in Habitica which would actually require it
directly and running the test suites without it didn't change anything.

To avoid adding yet another patch, I left it in since it's required for
babel/pug anyway. I did however ask the author of the change introducing
the dependency on GitHub whether this was intended.

Signed-off-by: aszlig <aszlig@nix.build>