Move Nix files around Node dependencies into deps/
First of all deps/generated makes a bit more sense in terms of what the
directory is all about and second, we want to make sure that overrides
and other helpers (eg. Google fonts, which also need Node packages) are
namespaced together with everything related to those dependencies.

And of course, this means three files less floating around in the
top-level directory.

Signed-off-by: aszlig <aszlig@nix.build>