Move docinfo.nix into pkgs/shabitica/
While I'm still not sure on the best location for this file, since it's
used by the main package, the modules and the tests, putting them
alongside the actual package has the advantage that we can someday just
pass through the attributes via the main Shabitica derivation.

This marks the last file I wanted to get off the top-level directory and
while there are still a few ugly bits here and there, especially
regarding package imports, I'm pretty much content for now when looking
at the source root.

Signed-off-by: aszlig <aszlig@nix.build>