Drop support for NixOS 18.03
NixOS 19.09 is coming up soon and 18.03 is no longer supported upstream
since ages, so it's time for us to get rid of all the workarounds we
needed to make just to support NixOS 18.03.

In the end, even NixOS 18.09 doesn't have upstream support anymore, but
it at least gets the most urgent security updates.

Additionally, we have currently only one(!) workaround for NixOS 18.09,
but a bunch of them for 18.03, so this is another reason why I want it
to be dropped ASAP, which by this change it is.

Signed-off-by: aszlig <aszlig@nix.build>