tests/e2e: Don't fail build with output
The E2E tests are flaky sometimes and the nightwatch HTML reports we're
getting are pretty much useless and if we fail with output the only way
to restart the test is by manually removing the output path.

So when failing without producing a store path the build output we're
getting is way more useful than the HTML report, so there is no reason
to make life more complicated for this.

Signed-off-by: aszlig <aszlig@nix.build>