tests/imageproxy: Move snakeoil cert to repository
The test sometimes fails when the derivation is built on different build
hosts, resulting in a certificate mismatch between the sslhost node and
the client node.

So to make sure we don't run into those failures again, let's just use a
hardcoded certificate and key, which also avoids the pair to be rebuilt
whenever we have an OpenSSL upgrade.

The pair was created via the following Nix expression:

  with import <nixpkgs> {};

  runCommand "snakeoil-keys" {
    nativeBuildInputs = [ openssl ];
    OPENSSL_CONF = writeText "snakeoil.cnf" ''
      [req]
      default_bits = 4096
      prompt = no
      default_md = sha256
      req_extensions = req_ext
      distinguished_name = dn
      [dn]
      CN = sslhost.org
      [req_ext]
      subjectAltName = DNS:sslhost.org
    '';
  } ''
    mkdir "$out"
    openssl req -x509 -newkey rsa:4096 -nodes -keyout "$out/key.pem" \
      -out "$out/cert.pem" -days 365000
  '';

Signed-off-by: aszlig <aszlig@nix.build>