tests/e2e: Rewrite test suite implementation
Since the upstream project doesn't even run those tests and they only
have a dummy test that is not even working, they pretty much are
eventually going to get rid of the E2E suite entirely.

In addition to that, I'm constantly annoyed at the lack of helpful
information the Nightwatch HTML report provides.

We already added a few more tests other than just the basic one in the
upstream project (which isn't working anyway), but writing tests via
patching Habitica is always a bit of a tedious task.

So I decided it would be better if we directly manage our own E2E tests
instead of using the broken ones from upstream.

The tests are implemented using pytest-selenium and use pytest-html to
render the reports along with screenshots and the HTML output of the
current test state.

Another improvement with the new implementation is that we now run the
tests locally instead of within the guest VM, so we can run the tests in
parallel. Of course, since we currently only have 3 tests, this doesn't
yet play out very well, but as soon as we expand on that it should make
a difference.

Communication between the test runner and the guest VM can now also be
done (although not used yet) via the run-in-guest helper, which
serialises arbitirary commands, runs them in the guest VM and sends back
the output.

Right now, the tests are more or less literally translated from the
JS-based tests, so they still have rough edges and a few things could be
refactored but they're already less flaky, mainly because we now re-run
tests if for example the WebDriver fails to initialise properly.

The latter unfortunately happens very frequently and since the fault is
not on our end, we don't have a better option rather than simply
re-running the tests.

With this change, we now bring back build failures with output which we
removed in c4f7970c4b748c4f2c39fb9425ff5a88fa15082d, so that we get a
nice HTML report on Hydra even on failure.

Signed-off-by: aszlig <aszlig@nix.build>