tests/e2e: Don't run in invite-only mode
By default, Shabitica is invite-only, but that makes it difficult to
create arbitrary test users. Fortunately, right now we only create two
users, which works when tests are run in parallel and both users are
created at roughly the same time.

On the flip side however, this will also cause random test failures,
because if one of those tests is quicker than the other, we'll get an
error during registration.

We'd have run into troubles anyway once we add more users, but for
checking whether invite-only mode works we have a dedicated test, so we
can disable it for E2E, which will also make test user registration much
simpler.

Signed-off-by: aszlig <aszlig@nix.build>