habitipy: Add setuptools to propagatedBuildInputs
In nixpkgs master, setuptools is no longer a runtime dependency by
default, but habitipy uses pkg_resources to query the path to
apidoc.txt.

Of course, when using an older version of nixpkgs, we implicitly get
setuptools and just adding it will add the dependency twice. However,
this doesn't hurt and it's better than adding a version constraints in
our Nix expressions.

Signed-off-by: aszlig <aszlig@nix.build>