Update to version 4.116.4
Upstream changes for version 4.116.4:

  * Fix: Adjust handling of "true" value in admin Hall of Heroes route
         so item paths are set correctly
  * Fix: If user has negative count of quest scrolls, correct to 0
         before making Gem purchase
  * Chore: Locale files updated
  * Chore: Add some more push notification types to user schema
  * Fix: Correct styling of usernames in profile links
  * Fix: Correct more font color styling in quest info popovers
  * Fix: Improve Record Yesterday's Activity modal styling

Upstream changes for version v4.116.3:

  * Content: Enable Supernatural Skins and Haunted Hair Colors
  * Chore: Bailey news announcement re the above
  * Feature: Add push notification for party activity

When rebasing the patches, there were a few merge conflicts, which
happened because of the push notification changes. There are now push
notifications for chat and a few more notification types, namely
mentionParty, mentionJoinedGuild and mentionUnjoinedGuild.

Of course, all of them are patched out accordingly on our side, because
they either use iOS or Android push services.

Signed-off-by: aszlig <aszlig@nix.build>