imageproxy: Allow nginx user to write to cache dir
Newer nginx versions in NixOS >= 20.09 versions seem to be more picky
about permissions for the cache directory and it also turned out that
even before NixOS 20.09 the caching wasn't working properly.

So we now not only allow the nginx user to write to that cache
directory, but also read its index, the latter which is used to
invalidate old cache entries.

Signed-off-by: aszlig <aszlig@nix.build>