imageproxy: Add cache dir to nginx ReadWritePaths
Since NixOS 20.09, nginx is using the ProtectSystem setting, which
conflicts somewhat with the systemd-confinement NixOS module I'm using
for my deployments (which is even more strict than the ProtectSystem
being set to "strict").

Fortunately while the ProtectSystem is in conflict, the ReadWritePaths
works fine, even for the NixOS confinement module, where it results into
a no-op.

The reason why I'm not adding another CacheDirectory to nginx is because
there is already a CacheDirectory defined, so simply using list merging
semantics of the module system doesn't work and second we want the
directory to be clearly detached from the nginx module, giving us more
control about its permissions and behaviour.

Signed-off-by: aszlig <aszlig@nix.build>