modules: Switch to NixOS systemd confinement
The origins of the (now upstreamed) systemd-confinement module were
actually here in Shabitica, back then when I was last working on the
project, I decided to still keep the ad-hoc implementation for
backwards-compatibility reasons.

However, since we're now already on NixOS 20.09 and the oldest release
we support is 20.03, there really is no reason anymore to *not* use the
upstream module.

Signed-off-by: aszlig <aszlig@nix.build>