mailer: Fix build on both NixOS 20.03 and 20.09
Unfortunately, we can't simply rely on the default "haskellPackages" set
to be sufficient for compiling on both releases, because NixOS 20.03
comes with network 2.x by default while 20.09 refers to network 3.x.

The issue in our case is that we used the fdSocket function, which was a
"pure" function in network 2.x but became an IO function in 3.x.

Instead of introducing weird hacks matching on specific versions, let's
for the time being just enforce GHC 8.8.4 for both NixOS releases, where
we now have network 3.x even for NixOS 20.03.

Of course, one of the downsides of this is that for NixOS 20.03 the
dependencies are not in the official NixOS binary cache, so people
deploying on 20.03 will probably face more things needed to be built.

Given that 20.03 is the older release, I think the solution is better
than having more to built for the newer release or even introduce more
wuird hacks (eg. patching "network").

In the long term however we eventually want Shabitica to be delivered as
a Nix Flake, where we can explicitly pin a specific version of nixpkgs
and thus won't have this problem any longer.

Signed-off-by: aszlig <aszlig@nix.build>