imageproxy: Add a few more allowed syscalls
This is something we should investigate, since oddly on NixOS 20.09
tests run fine without these syscalls allowed but NixOS 20.03 *and*
NixOS Unstable panic.

Fortunately, those additional syscalls don't increase the attack surface
by quite a lot, except of course ioctl(). I tried to avoid it, but
apparently Go is using it in conjunction with FIONREAD to determine the
size of the input buffer. Ideally, we'd just add a small match on that
flag and only allow this specific ioctl, but unfortunately, systemd only
allows very simple syscall filtering.

Signed-off-by: aszlig <aszlig@nix.build>