tests/e2e: Don't import python-packages.nix
Python packages from nixpkgs should already be available via
pythonXPackages or pythonX.pkgs, so we should not really need to import
the package set directly.

The reason why this is necessary is because in NixOS 21.05 the
python-packages.nix expression requires an additional "self" argument.

So using pkgs.python37.pkgs instead should hopefully be
backwards-compatible as well.

Signed-off-by: aszlig <aszlig@nix.build>