tests/backup: Fix build with NixOS 21.05
Since NixOS 21.05 uses flake8 instead of black, we now get errors
whenever we 're using an f-string without placeholders. In this test
however, we're using Nix as a macro language which also *allows* (and
uses depending on function call usage) f-string placeholder.

To get around that, let's simply place a "noqa" comment after the
f-string in question to ignore the error.

Signed-off-by: aszlig <aszlig@nix.build>