tests/backup: Fix build with NixOS 21.05 for real
This reverts commit b349d4e475c03f24f83fafcf0e8bfe023ad8c739.

The commit in question added a "# noqa" comment to the end of the line
where we don't care about whether an f-string doesn't contain a
placeholder.

Unfortunately, "# noqa" is only allowed in Flake8, but not in Pyflakes
so to work around this I added a dummy placeholder with an empty string
to shut up Pyflakes.

Signed-off-by: aszlig <aszlig@nix.build>